requirements_lambda.txt
fastapi==0.104.1 
mangum==0.17.0 
numpy==1.24.3 
pillow==10.1.0 
python-multipart==0.0.6 
